dev/workAutomation/requirements.txt
djangorestframework
markdown
django-filter
gunicorn
pandas
openpyxl
pymysql